refactor: relocate events (API)
