+++ src/api/events/property-inspector.ts
