+++ src/events/send-to-plugin-event.ts
@@ -1,5 +1,5 @@
-import { Action } from "../actions/action";
-import type { PayloadObject, SendToPlugin } from "../connectivity/events";
+import type { Action } from "../actions/action";
+import type { PayloadObject, SendToPlugin } from "../api/events";
 import { Event } from "./event";
 
 /**
