+++ src/system.ts
@@ -1,6 +1,6 @@
+import type { SystemDidWakeUp } from "./api/events";
 import type { IDisposable } from "./common/disposable";
 import type { StreamDeckConnection } from "./connectivity/connection";
-import type { SystemDidWakeUp } from "./connectivity/events";
 import { ApplicationDidLaunchEvent, ApplicationDidTerminateEvent, ApplicationEvent, DidReceiveDeepLinkEvent, Event, SystemDidWakeUpEvent } from "./events";
 import type { Manifest } from "./manifest";
 import { requiresVersion } from "./validation";
